+++ pitest/src/main/java/org/pitest/coverage/CoverageReceiver.java
@@ -7,10 +7,10 @@ import sun.pitest.InvokeReceiver;
 public interface CoverageReceiver extends InvokeReceiver {
 
   void recordTestOutcome(Description description, boolean wasGreen,
-      int executionTime);
+      double executionTime);
 
   default void recordTestOutcome(Description description, boolean wasGreen,
-      int executionTime, String exceptionType, String exceptionMessage, String stackTrace) {
+      double executionTime, String exceptionType, String exceptionMessage, String stackTrace) {
     // Default implementation for backward compatibility
     recordTestOutcome(description, wasGreen, executionTime);
   }
